Minor restructuring and renaming
